+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 * removals
   * `Devise::Mailer#scope_name` and `Devise::Mailer#resource` are now protected
     methods instead of public.
+* enhancements
+  * Notify the original email when it is changed with a new `Devise.send_email_changed_notification` setting.
+    When using `reconfirmable`, the notification will be sent right away instead of when the unconfirmed email is confirmed.
+    (original change by @ethirajsrinivasan)
 
 ### 4.2.0 - 2016-07-01